backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
transformers==4.11.3
torch==1.9.0
httpx==0.19.0 